Edited build properties
Re-edited the build properties such that the library actuaully builds when I want it to build. I was able to get it to successfully build today, because the path for the Processing app was broken. It actually had to go all the way into my hard drive folder, and get it from the total applications folder. I never realized that there were two before.