New updates: new objects
Two new objects introduced with this update: Row and ObjectLabel. Row is currently marked as public but that is only for my own benefit as I am debugging this. Row will ultimately be used by matrix to sort various things into their own rows or columns. Exciting stuff.
ObjectLabel is merely a Label but with an object stored inside. Therefore, it can dynamically keep track of the location of the object without needing any further user input.