Tested on Processing 2.2.1
Opened up my overall-testing sketch file on Processing 2.2.1 and it worked completely. This is probably because I don't do anything fancy like threading and because I built this in 3.0.2, my library doesn't assume that PApplet is a subclass of Applet. This might actually make testing a lot easier than I expected. However, I'm not exactly sure how I indicate that this is compatible with both.