Housecleaning/Update
Deleted the Nothing class because it was just supposed to be a test anyway.
NumOutput now limits the number of digits you'll have trailing after the number.
New class, Point: stores an x and a y. This is helpful for many applications where making global x and y variables simply isn't viable.
RoundedRectButton is now fully working, again.