New updates for Frame and RoundedRectButton
I've been trying to code a version of TextTwist so I'm using the RoundButtons as the bubbles for letters. To move them around, I just put in a new function in Frame, RoundButton's superclass, that lets you change its x and y positions accordingly. I believe this update also includes changes to the way thta RoundedRectButton is called: Apparently there is a way to make rounded rectangles using rect() and I never knew about it until like two days ago, when I clicked on the 'rect' link on the Processing reference page accidentally